Add failing test for missing events
Discovered a case where while the connector is down, events published to the wal will be
skipped upon restart of the connector. This is due to the starting LSN getting set to the
xlog position for the transaction that starts the snapshot phase. The existing "always"
snapshotter relies on this behavior to avoid duplicate messages. Fixing this requires
a patch for Debezium so development will move to the latest sources and work from there.