Remove reliance on JMX metrics
JMX metrics are flaky to use and are not always available.
For example, GraalVM does not currently support JMX. Determining
when a connector is finished snapshotting is still tricky because we need
to rely on counting calls on the snapshotter interface for methods that are
called after the snapshot is known to finish. The function shouldStream() is
called during the streaming phase and initalization. This is still very
hacky but is still preferrable to relying on JMX. Downsides here are that
we need to carefully watch the Debezium code at releases to make sure the
call counts remain accurate. Long term, a close handler added to the
snapshotter interface would be a nice addition added upstream to Debezium.