Add support for future 1.4 debezium support
The 1.4 release brings in a new lifecycle hook Snapshotter#snapshotCompleted
that will significantly simplify our the complexity around pg connection
management.  Previously, the partial snapshotter needed to use JMX to determine
when the snapshot was finished and as a result, the snapshot filter needed
to run in a separate thread. With the new lifecycle hook, the filter removes
the reliance on JMX and doesn't need the extra thread.